Data.Dequeue.Show

Description

A newtype used entirely for its derived Read and Show instances. These are then used by showDequeue and readDequeue to make writing Read and Show instances for Dequeues easier.

Documentation

newtype Dequeue a

Constructors

Dequeue [a]